refactor(app): extract shared tier price fields component
The interval/amount/currency fieldset was duplicated near
character-for-character between TierPriceForm and TierForm's
initial-price rows. Move it (and intervalLabels) into a generic
TierPriceFields component keyed by caller-supplied field paths, and drop
the defensive optional chaining on RHF onChange handlers along the way.
